refactor(primitives): move divider to primitives
