+++ src/components/primitives/divider/index.ts
