+++ src/components/primitives/divider/types.ts
